database and create account intergration
I spent majority of the past week working on the create account window and linking it to my passwordmanager database. I made the UI and made sure that the data entered by the user was saved in the database. This allowed the user access to the central command window but I think this name will change but am not sure. When I tested whether it worked, I got the appropriate result, everything worked fine. I also added new functions to the functions window and a tst.py file but I am gonna delete that file right after this becasue I have already added it